+++ axum/src/extract/ws.rs
@@ -267,14 +267,15 @@ where
             return Err(InvalidWebSocketVersionHeader.into());
         }
 
-        let sec_websocket_key =
-            if let Some(key) = req.headers_mut().remove(header::SEC_WEBSOCKET_KEY) {
-                key
-            } else {
-                return Err(WebSocketKeyHeaderMissing.into());
-            };
+        let sec_websocket_key = req
+            .headers_mut()
+            .remove(header::SEC_WEBSOCKET_KEY)
+            .ok_or(WebSocketKeyHeaderMissing)?;
 
-        let on_upgrade = req.extensions_mut().remove::<OnUpgrade>().unwrap();
+        let on_upgrade = req
+            .extensions_mut()
+            .remove::<OnUpgrade>()
+            .ok_or(MissingUpgradeState)?;
 
         let sec_websocket_protocol = req.headers().get(header::SEC_WEBSOCKET_PROTOCOL).cloned();
 
@@ -564,6 +565,16 @@ pub mod rejection {
         pub struct WebSocketKeyHeaderMissing;
     }
 
+    define_rejection! {
+        #[status = BAD_REQUEST]
+        #[body = "WebSocket request couldn't be upgraded since no upgrade state was present"]
+        /// Rejection type for [`WebSocketUpgrade`](super::WebSocketUpgrade).
+        ///
+        /// See <https://docs.rs/hyper/latest/hyper/upgrade/index.html> for more details on how
+        /// hyper handles HTTP upgrades.
+        pub struct MissingUpgradeState;
+    }
+
     composite_rejection! {
         /// Rejection used for [`WebSocketUpgrade`](super::WebSocketUpgrade).
         ///
@@ -575,6 +586,7 @@ pub mod rejection {
             InvalidUpgradeHeader,
             InvalidWebSocketVersionHeader,
             WebSocketKeyHeaderMissing,
+            MissingUpgradeState,
         }
     }
 }
@@ -639,4 +651,37 @@ pub mod close_code {
     /// IP (when multiple targets exist), or reconnect to the same IP when a user has performed an
     /// action.
     pub const AGAIN: u16 = 1013;
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::{body::Body, routing::get};
+    use http::Request;
+    use tower::ServiceExt;
+
+    #[tokio::test]
+    async fn rejects_requests_without_upgrade() {
+        let svc = get(|ws: Result<WebSocketUpgrade, WebSocketUpgradeRejection>| {
+            let rejection = ws.unwrap_err();
+            assert!(matches!(
+                rejection,
+                WebSocketUpgradeRejection::MissingUpgradeState(_)
+            ));
+            std::future::ready(())
+        });
+
+        let req = Request::builder()
+            .method(Method::GET)
+            .header("upgrade", "websocket")
+            .header("connection", "Upgrade")
+            .header("sec-websocket-key", "6D69KGBOr4Re+Nj6zx9aQA==")
+            .header("sec-websocket-version", "13")
+            .body(Body::empty())
+            .unwrap();
+
+        let res = svc.oneshot(req).await.unwrap();
+
+        assert_eq!(res.status(), StatusCode::OK);
+    }
 }